Correction bug : Problème de type : passage de float a double

+++ src/general/Constantes.java
@@ -11,7 +11,7 @@ import javax.swing.JOptionPane;
 
 public class Constantes {
 	public final static int INTEGER = 1;
-	public final static int FLOAT = 2;
+	public final static int DOUBLE = 2;
 	public final static int BOOLEAN = 3;
 	public final static int STRING = 4;
 	public final static Object[] options = {"Annuler", "Non", "Oui"};
@@ -24,7 +24,7 @@ public class Constantes {
 	public static int stringToInt (String st) {
 		int res = 0;
 		if (st.equals("integer")) res = INTEGER;
-		else if (st.equals("float")) res = FLOAT;
+		else if (st.equals("float")) res = DOUBLE;
 		else if (st.equals("text")) res = STRING;
 		else if (st.equals("bool")) res = BOOLEAN;
 		return res;
